data/person_background_cutmix/cutout_tool/requirements_cutout.txt
rembg[cpu]>=2.0.75
Pillow>=10.0.0
